Fjernet gammel footer
